repository: merge check.go into checker.go
